+++ acir/src/native_types/witness.rs
@@ -6,6 +6,7 @@ use flate2::{
 };
 use serde::{Deserialize, Serialize};
 
+// Witness might be a misnomer. This is an index that represents the position a witness will take
 #[derive(
     Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Default, Serialize, Deserialize,
 )]
@@ -55,7 +56,7 @@ impl Witness {
 // We use this, so that they are pushed to the beginning of the array
 //
 // When they are pushed to the beginning of the array, they are less likely to be used in an intermediate gate
-// by the optimiser, which would mean two unknowns in an equation.
+// by the optimizer, which would mean two unknowns in an equation.
 // See Issue #20
 // TODO: can we find a better solution to this?
 pub struct UnknownWitness(pub u32);