+++ acvm/src/compiler/optimiser/mod.rs
@@ -1,7 +1,0 @@
-mod csat_optimiser;
-mod general_optimiser;
-mod r1cs_optimiser;
-
-pub use csat_optimiser::Optimiser as CSatOptimiser;
-pub use general_optimiser::GeneralOpt as GeneralOptimiser;
-pub use r1cs_optimiser::R1CSOptimiser;